pyproject.toml
[project]
name = "council"
version = "3.0.0"
description = "Multi-model planning + voice-controlled agent dispatcher"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.24",
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.scripts]
council = "council.cli:cli"
dispatcher = "council.dispatcher.simple:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
